fix: products use cases exception handling

+++ src/modules/products/use-cases/delete-products/delete-products.use-case.ts
@@ -1,3 +1,4 @@
+import NotFoundError from '@errors/not-found.error'
 import { IUseCase } from '@interfaces/use-case.interface'
 import { IProductsRepository } from '@modules/products/repositories/products-repository.interface'
 
@@ -7,7 +8,8 @@ export class DeleteProductsUseCase implements IUseCase {
   ) {}
 
   async execute (id: number): Promise<string> {
-    const response = await this.productsRepository.getOne({ id, throws: true })
+    const response = await this.productsRepository.getOne({ id })
+    if (!response) throw new NotFoundError('Product not found')
     await this.productsRepository.remove(response)
     return this.productsRepository.getDeleteMessage()
   }
